Revert "fix methods to get all printers and default printer (#121)"
This reverts commit e729c330e5c99b69716ee2bc070321fa3c6d076c.
