--- src/unix/get-default-printer.js
@@ -1,10 +1,0 @@
-"use strict";
-
-const execAsync = require("../execAsync");
-
-const getDefaultPrinter = () => {
-  const parseResult = output => output.trim().substr(output.indexOf(":") + 1);
-  return execAsync("lpstat", ["-d"], parseResult);
-};
-
-module.exports = getDefaultPrinter;
