+++ src/unix/index.js
@@ -1,3 +1,2 @@
-export { default as getDefaultPrinter } from "./get-default-printer";
-export { default as getPrinters } from "./get-printers";
+export * from "./printers";
 export { default as print } from "./print";
