+++ src/unix/printers.js
@@ -12,4 +12,12 @@ const getPrinters = () => {
   return execAsync("lpstat", ["-a"], parseResult);
 };
 
-module.exports = getPrinters;
+const getDefaultPrinter = () => {
+  const parseResult = output => output.substr(output.indexOf(":") + 1).trim();
+  return execAsync("lpstat", ["-d"], parseResult);
+};
+
+module.exports = {
+  getPrinters,
+  getDefaultPrinter
+};
